+++ src/openvpn/Makefile.am
@@ -108,6 +108,7 @@ openvpn_SOURCES = \
 	pkcs11_mbedtls.c \
 	openvpn.c openvpn.h \
 	options.c options.h \
+	options_show.c options_show.h \
 	options_util.c options_util.h \
 	options_parse.c \
 	otime.c otime.h \
