+++ src/openvpn/openvpn.c
@@ -30,6 +30,7 @@
 #include "forward.h"
 #include "multi.h"
 #include "win32.h"
+#include "options_show.h"
 #include "platform.h"
 #include "string.h"
